+++ Sources/BugsnagPerformance/Private/RetryQueue.mm
@@ -44,11 +44,15 @@ void RetryQueue::configure(BugsnagPerformanceConfiguration *config) noexcept {
 }
 
 void RetryQueue::preStartSetup() noexcept {
-    [Filesystem ensurePathExists:baseDir_];
+    ensureBaseDirExists();
 }
 
 void RetryQueue::sweep() noexcept {
     ensureBaseDirExists();
+    if (storageDisabled_.load(std::memory_order_acquire)) {
+        return;
+    }
+
     NSError *error = nil;
     auto contents = [[NSFileManager defaultManager] contentsOfDirectoryAtPath:baseDir_ error:&error];
     if (contents == nil) {
@@ -71,6 +75,10 @@ void RetryQueue::sweep() noexcept {
 
 std::vector<dispatch_time_t> RetryQueue::list() noexcept {
     ensureBaseDirExists();
+    if (storageDisabled_.load(std::memory_order_acquire)) {
+        return {};
+    }
+
     NSError *error = nil;
     std::vector<dispatch_time_t> timestamps;
     auto contents = [[NSFileManager defaultManager] contentsOfDirectoryAtPath:baseDir_ error:&error];
@@ -93,6 +101,11 @@ std::vector<dispatch_time_t> RetryQueue::list() noexcept {
 }
 
 std::unique_ptr<OtlpPackage> RetryQueue::get(dispatch_time_t ts) noexcept {
+    ensureBaseDirExists();
+    if (storageDisabled_.load(std::memory_order_acquire)) {
+        return nullptr;
+    }
+
     // If this fails, we don't care why. Just delete the file.
     NSData *contents = [NSData dataWithContentsOfFile:fullPath(filenameFromTimestamp(ts))
                                               options:0 error:nil];
@@ -112,6 +125,10 @@ std::unique_ptr<OtlpPackage> RetryQueue::get(dispatch_time_t ts) noexcept {
 
 void RetryQueue::add(OtlpPackage &package) noexcept {
     ensureBaseDirExists();
+    if (storageDisabled_.load(std::memory_order_acquire)) {
+        return;
+    }
+
     auto data = package.serialize();
     NSError *error = nil;
     NSString *filePath = fullPath(filenameFromTimestamp(package.timestamp));
@@ -122,6 +139,9 @@ void RetryQueue::add(OtlpPackage &package) noexcept {
 }
 
 void RetryQueue::remove(dispatch_time_t ts) noexcept {
+    if (storageDisabled_.load(std::memory_order_acquire)) {
+        return;
+    }
     remove(filenameFromTimestamp(ts));
 }
 
@@ -135,9 +155,31 @@ NSString * RetryQueue::fullPath(NSString *filename) noexcept {
 }
 
 void RetryQueue::ensureBaseDirExists() noexcept {
-    NSError *error = [Filesystem ensurePathExists:baseDir_];
-    if (error != nil) {
-        BSGLogError(@"error while creating base dir %@: %@", baseDir_, error);
-        onFilesystemError();
+    if (storageDisabled_.load(std::memory_order_acquire)) {
+        return;
     }
+
+    NSError *error = nil;
+    if (ensurePathFunc_) {
+        error = ensurePathFunc_(baseDir_);
+    } else {
+        error = [Filesystem ensurePathExists:baseDir_];
+    }
+
+    if (error != nil) {
+        storageDisabled_.store(true, std::memory_order_release);
+
+        // notify once, thread-safely
+        const bool alreadyNotified =
+            didNotifyStorageError_.exchange(true, std::memory_order_acq_rel);
+
+        if (!alreadyNotified) {
+            BSGLogError(@"error while creating base dir %@: %@", baseDir_, error);
+            onFilesystemError();
+        }
+    }
+}
+
+void RetryQueue::setEnsurePathExistsHandler(std::function<NSError *(NSString *)> handler) noexcept {
+    ensurePathFunc_ = handler;
 }