+++ CHANGELOG.md
@@ -3,7 +3,10 @@ Changelog
 
 ## TBD
 
-### Bug fixes
+### Bug Fixes
+
+* Don't attempt file operations when top-level directory creation failed
+  [570](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/570)
 
 * Support Pascal case plist configuration values.
   [#571](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/571)